# walletnode/api/gen/http/openapi3.yaml
openapi: 3.0.3
info:
  title: WalletNode REST API
  version: "1.0"
servers:
- url: http://localhost:8080
paths:
...
  /nfts/download:
    get:
      tags:
      - nft
      summary: Downloads NFT
      description: Download registered NFT.
      operationId: nft#download
      parameters:
      - name: txid
        in: query
        description: Nft Registration Request transaction ID
        allowEmptyValue: true
        required: true
        schema:
          type: string
          description: Nft Registration Request transaction ID
          example: 576e7b824634a488a2f0baacf5a53b237d883029f205df25b300b87c8877ab58
          minLength: 64
          maxLength: 64
        example: 576e7b824634a488a2f0baacf5a53b237d883029f205df25b300b87c8877ab58
      - name: pid
        in: query
        description: Owner's PastelID
        allowEmptyValue: true
        required: true
        schema:
          type: string
          description: Owner's PastelID
          example: jXYJud3rmrR1Sk2scvR47N4E4J5Vv48uCC6se2nzHrBRdjaKj3ybPoi1Y2VVoRqi1GnQrYKjSxQAC7NBtvtEdS
          pattern: ^[a-zA-Z0-9]+$
          minLength: 86
          maxLength: 86
        example: jXYJud3rmrR1Sk2scvR47N4E4J5Vv48uCC6se2nzHrBRdjaKj3ybPoi1Y2VVoRqi1GnQrYKjSxQAC7NBtvtEdS
      responses:
        "200":
          description: OK response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadResult'
              example:
                file:
                - 81
                - 117
                - 105
                - 32
                - 113
                - 117
                - 111
                - 32
                - 101
                - 116
                - 32
                - 111
                - 100
                - 105
                - 111
                - 32
                - 118
                - 111
                - 108
                - 117
                - 112
                - 116
                - 97
                - 116
                - 101
                - 109
                - 32
                - 118
                - 111
                - 108
                - 117
                - 112
                - 116
                - 97
                - 115
                - 32
                - 118
                - 111
                - 108
                - 117
                - 112
                - 116
                - 97
                - 116
                - 101
                - 115
                - 46
        "404":
          description: Not Found response.
          content:
            application/vnd.goa.error:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                id: 3F1FKVRR
                message: Value of ID must be an integer
                name: bad_request
        "500":
          description: Internal Server Error response.
          content:
            application/vnd.goa.error:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                id: 3F1FKVRR
                message: Value of ID must be an integer
                name: bad_request
      security:
      - api_key_header_Authorization: []
...
  /nfts/register/{taskId}/state:
    get:
      tags:
      - nft
      summary: Streams state by task ID
      description: Streams the state of the registration process.
      operationId: nft#registerTaskState
      parameters:
      - name: taskId
        in: path
        description: Task ID of the registration process
        required: true
        schema:
          type: string
          description: Task ID of the registration process
          example: n6Qn6TFM
          minLength: 8
          maxLength: 8
        example: n6Qn6TFM
      responses:
        "101":
          description: Switching Protocols response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskState'
              example:
                date: 2006-01-02T15:04:05Z07:00
                status: Task Started
        "404":
          description: Not Found response.
          content:
            application/vnd.goa.error:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                id: 3F1FKVRR
                message: Value of ID must be an integer
                name: bad_request
        "500":
          description: Internal Server Error response.
          content:
            application/vnd.goa.error:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                id: 3F1FKVRR
                message: Value of ID must be an integer
                name: bad_request
  /nfts/register/upload:
    post:
      tags:
      - nft
      summary: Uploads an image
      description: Upload the image that is used when registering a new NFT.
      operationId: nft#uploadImage
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UploadImageRequestBody'
            example:
              file:
              - 73
              - 112
              - 115
              - 117
              - 109
              - 32
              - 116
              - 111
              - 116
              - 97
              - 109
              - 32
              - 101
              - 115
              - 116
              - 32
              - 117
              - 110
              - 100
              - 101
              - 32
              - 113
              - 117
              - 105
              - 97
              - 46
      responses:
        "201":
          description: Created response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Image'
              example:
                expires_in: 2006-01-02T15:04:05Z07:00
                image_id: VK7mpAqZ
        "400":
          description: Bad Request response.
          content:
            application/vnd.goa.error:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                id: 3F1FKVRR
                message: Value of ID must be an integer
                name: bad_request
        "500":
          description: Internal Server Error response.
          content:
            application/vnd.goa.error:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                id: 3F1FKVRR
                message: Value of ID must be an integer
                name: bad_request
...
  /openapi/cascade/details/{image_id}:
    post:
      tags:
      - cascade
      summary: Provides action details
      description: Provide action details
      operationId: cascade#actionDetails
      parameters:
      - name: image_id
        in: path
        description: Uploaded image ID
        required: true
        schema:
          type: string
          description: Uploaded image ID
          example: VK7mpAqZ
          minLength: 8
          maxLength: 8
        example: VK7mpAqZ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActionDetailsRequestBody'
            example:
              action_data_hash: 7ae3874ff2df92df38cce7586c08fe8f3687884edf3b0543f8d9420f4df31265
              action_data_signature: bTwvO6UZSvFqHb9qmXbAOg2VmupmP70wfhYsAvwFfeC61cuoL9KIXZtbdQ/Ek8FVNoTpCY5BuxcA6lNjkIOBh4w9/RWtuqF16IaJhAnZ4JbZm1MiDCGcf7x0UU/GNRSk6rNAHlPYkOPdhkha+JjCwD4A
              app_pastelid: jXZqaS48TT6LFjxnf9P68hZNQVCFqY631FPz4CtM6VugDi5yLNB51ccy17kgKKCyFuL4qadQsXHH3QwHVuPVyY
      responses:
        "201":
          description: Created response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionDetailResult'
              example:
                estimated_fee: 100
        "400":
          description: Bad Request response.
          content:
            application/vnd.goa.error:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                id: 3F1FKVRR
                message: Value of ID must be an integer
                name: bad_request
        "500":
          description: Internal Server Error response.
          content:
            application/vnd.goa.error:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                id: 3F1FKVRR
                message: Value of ID must be an integer
                name: bad_request
  /openapi/cascade/download:
    get:
      tags:
      - cascade
      summary: Downloads cascade artifact
      description: Download cascade Artifact.
      operationId: cascade#download
      parameters:
      - name: txid
        in: query
        description: Nft Registration Request transaction ID
        allowEmptyValue: true
        required: true
        schema:
          type: string
          description: Nft Registration Request transaction ID
          example: 576e7b824634a488a2f0baacf5a53b237d883029f205df25b300b87c8877ab58
          minLength: 64
          maxLength: 64
        example: 576e7b824634a488a2f0baacf5a53b237d883029f205df25b300b87c8877ab58
      - name: pid
        in: query
        description: Owner's PastelID
        allowEmptyValue: true
        required: true
        schema:
          type: string
          description: Owner's PastelID
          example: jXYJud3rmrR1Sk2scvR47N4E4J5Vv48uCC6se2nzHrBRdjaKj3ybPoi1Y2VVoRqi1GnQrYKjSxQAC7NBtvtEdS
          pattern: ^[a-zA-Z0-9]+$
          minLength: 86
          maxLength: 86
        example: jXYJud3rmrR1Sk2scvR47N4E4J5Vv48uCC6se2nzHrBRdjaKj3ybPoi1Y2VVoRqi1GnQrYKjSxQAC7NBtvtEdS
      responses:
        "200":
          description: OK response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadResult'
              example:
                file:
                - 65
                - 117
                - 116
                - 32
                - 117
                - 116
                - 32
                - 118
                - 111
                - 108
                - 117
                - 112
                - 116
                - 97
                - 116
                - 101
                - 115
                - 32
                - 114
                - 101
                - 114
                - 117
                - 109
                - 46
        "404":
          description: Not Found response.
          content:
            application/vnd.goa.error:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                id: 3F1FKVRR
                message: Value of ID must be an integer
                name: bad_request
        "500":
          description: Internal Server Error response.
          content:
            application/vnd.goa.error:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                id: 3F1FKVRR
                message: Value of ID must be an integer
                name: bad_request
      security:
      - api_key_header_Authorization: []
  /openapi/cascade/start/{image_id}:
    post:
      tags:
      - cascade
      summary: Starts processing the image
      description: Start processing the image
      operationId: cascade#startProcessing
      parameters:
      - name: image_id
        in: path
        description: Uploaded image ID
        required: true
        schema:
          type: string
          description: Uploaded image ID
          example: VK7mpAqZ
          minLength: 8
          maxLength: 8
        example: VK7mpAqZ
      - name: app_pastelid_passphrase
        in: header
        description: Passphrase of the App PastelID
        allowEmptyValue: true
        required: true
        schema:
          type: string
          description: Passphrase of the App PastelID
          example: qwerasdf1234
        example: qwerasdf1234
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartProcessingRequestBody'
            example:
              app_pastelid: jXYJud3rmrR1Sk2scvR47N4E4J5Vv48uCC6se2nzHrBRdjaKj3ybPoi1Y2VVoRqi1GnQrYKjSxQAC7NBtvtEdS
              burn_txid: 576e7b824634a488a2f0baacf5a53b237d883029f205df25b300b87c8877ab58
      responses:
        "201":
          description: Created response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StartProcessingResult'
              example:
                task_id: VK7mpAqZ
        "400":
          description: Bad Request response.
          content:
            application/vnd.goa.error:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                id: 3F1FKVRR
                message: Value of ID must be an integer
                name: bad_request
        "500":
          description: Internal Server Error response.
          content:
            application/vnd.goa.error:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                id: 3F1FKVRR
                message: Value of ID must be an integer
                name: bad_request
  /openapi/cascade/start/{taskId}/history:
    get:
      tags:
      - cascade
      summary: Get history of states as a json string with a list of state objects.
      description: Gets the history of the task's states.
      operationId: cascade#getTaskHistory
      parameters:
      - name: taskId
        in: path
        description: Task ID of the registration process
        required: true
        schema:
          type: string
          description: Task ID of the registration process
          example: n6Qn6TFM
          minLength: 8
          maxLength: 8
        example: n6Qn6TFM
      responses:
        "200":
          description: OK response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskHistory'
              example:
                list: '[ Started, Image Probed, Downloaded... ]'
        "404":
          description: Not Found response.
          content:
            application/vnd.goa.error:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                id: 3F1FKVRR
                message: Value of ID must be an integer
                name: bad_request
        "500":
          description: Internal Server Error response.
          content:
            application/vnd.goa.error:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                id: 3F1FKVRR
                message: Value of ID must be an integer
                name: bad_request
...
components:
  schemas:
...
    DownloadResult:
      type: object
      properties:
        file:
          type: string
          description: File downloaded
          example:
          - 69
          - 120
          - 32
          - 112
          - 108
          - 97
          - 99
          - 101
          - 97
          - 116
          - 32
          - 111
          - 112
          - 116
          - 105
          - 111
          - 32
          - 113
          - 117
          - 97
          - 101
          - 114
          - 97
          - 116
          - 32
          - 110
          - 97
          - 109
          - 32
          - 114
          - 101
          - 112
          - 114
          - 101
          - 104
          - 101
          - 110
          - 100
          - 101
          - 114
          - 105
          - 116
          - 46
          format: binary
      example:
        file:
        - 78
        - 111
        - 110
        - 32
        - 97
        - 117
        - 116
        - 46
      required:
      - file
    Error:
      type: object
      properties:
        fault:
          type: boolean
          description: Is the error a server-side fault?
          example: false
        id:
          type: string
          description: ID is a unique identifier for this particular occurrence of
            the problem.
          example: 123abc
        message:
          type: string
          description: Message is a human-readable explanation specific to this occurrence
            of the problem.
          example: parameter 'p' must be an integer
        name:
          type: string
          description: Name is the name of this class of errors.
          example: bad_request
        temporary:
          type: boolean
          description: Is the error temporary?
          example: false
        timeout:
          type: boolean
          description: Is the error a timeout?
          example: true
      example:
        id: 3F1FKVRR
        message: Value of ID must be an integer
        name: bad_request
      required:
      - name
      - id
      - message
      - temporary
      - timeout
      - fault
...
  securitySchemes:
    api_key_header_Authorization:
      type: apiKey
      description: Nft Owner's passphrase to authenticate
      name: Authorization
      in: header
...